+++ nullplatform/cloud/aws/vpc/main.tf
@@ -3,18 +3,19 @@ resource "nullplatform_provider_config" "vpc" {
   nrn        = var.nrn
   type       = "aws-networking-configuration"
   dimensions = var.dimensions
-  attributes = jsonencode({
-    vpc = {
-      id              = var.vpc_id
-      subnets         = var.vpc_subnets
-      security_groups = var.vpc_security_groups
-    }
-    load_balancer = {
-      public  = var.load_balancer.public
-      private = var.load_balancer.private
-    }
-  })
-  lifecycle {
-    ignore_changes = [attributes]
-  }
+  attributes = jsonencode(merge(
+    {
+      vpc = {
+        id              = var.vpc_id
+        subnets         = var.vpc_subnets
+        security_groups = var.vpc_security_groups
+      }
+    },
+    local.load_balancer_configured ? {
+      load_balancer = {
+        public  = var.load_balancer.public
+        private = var.load_balancer.private
+      }
+    } : {}
+  ))
 }