+++ nullplatform/scope_configuration/main.tf
@@ -1,10 +1,7 @@
 resource "nullplatform_provider_config" "scope_configuration" {
   nrn        = var.nrn
-  type       = var.provider_specification_slug
+  type       = var.type
   dimensions = var.dimensions
-  attributes = jsonencode(var.attributes)
 
-  lifecycle {
-    ignore_changes = [attributes]
-  }
+  attributes = jsonencode(merge(local.defaults, local.overrides))
 }